Move credentials methods to pipeline service.
