+++ test/unit/services/base/service_test.py
@@ -1,8 +1,5 @@
 import io
-import json
-import jwt
 
-from amqpstorm import AMQPError
 from unittest.mock import patch
 from unittest.mock import call
 from unittest.mock import MagicMock, Mock
@@ -53,14 +50,8 @@ class TestBaseService(object):
         mock_get_configuration.return_value = config
 
         self.service = MashService('obs')
-        self.service.encryption_keys_file = 'encryption_keys.file'
-        self.service.jwt_secret = 'a-secret'
-        self.service.jwt_algorithm = 'HS256'
 
         mock_get_configuration.assert_called_once_with('obs')
-        config.get_encryption_keys_file.assert_called_once_with()
-        config.get_jwt_secret.assert_called_once_with()
-        config.get_jwt_algorithm.assert_called_once_with()
         mock_get_job_directory.assert_called_once_with('obs')
         mock_makedirs.assert_called_once_with(
             '/var/lib/mash/obs_jobs/', exist_ok=True
@@ -117,22 +108,6 @@ class TestBaseService(object):
         self.service.consume_queue(callback)
         self.channel.basic.consume.assert_called_once_with(
             callback=callback, queue='obs.service'
-        )
-
-    def test_consume_credentials_queue(self):
-        callback = Mock()
-
-        self.service.consume_credentials_queue(callback)
-        self.channel.basic.consume.assert_called_once_with(
-            callback=callback, queue='obs.credentials'
-        )
-
-    @patch.object(MashService, 'bind_queue')
-    def test_bind_credentials_queue(self, mock_bind_queue):
-        self.service.bind_credentials_queue()
-
-        mock_bind_queue.assert_called_once_with(
-            'obs', 'response', 'credentials'
         )
 
     def test_close_connection(self):
@@ -204,131 +179,6 @@ class TestBaseService(object):
         )
         self.service.channel.queue.unbind.assert_called_once_with(
             queue='testing.service', exchange='testing', routing_key='1'
-        )
-
-    def test_get_credentials_request(self):
-        self.service.jwt_algorithm = 'HS256'
-        self.service.jwt_secret = 'super.secret'
-        self.service_exchange = 'obs'
-        message = self.service.get_credential_request('1')
-
-        token = json.loads(message)['jwt_token']
-        payload = jwt.decode(
-            token, 'super.secret', algorithm='HS256',
-            issuer='obs', audience='credentials'
-        )
-
-        assert payload['id'] == '1'
-        assert payload['sub'] == 'credentials_request'
-
-    def test_get_encryption_keys_from_file(self):
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.readlines.return_value = [
-                '1234567890123456789012345678901234567890123=\n'
-            ]
-            result = self.service.get_encryption_keys_from_file(
-                'test-keys.file'
-            )
-
-        assert len(result) == 1
-        assert type(result[0]).__name__ == 'Fernet'
-
-    @patch.object(MashService, 'decrypt_credentials')
-    @patch('mash.services.mash_service.jwt')
-    def test_decode_credentials(self, mock_jwt, mock_decrypt):
-        self.service.jwt_algorithm = 'HS256'
-        self.service.jwt_secret = 'super.secret'
-        self.service_exchange = 'obs'
-
-        message = {'jwt_token': 'secret_credentials'}
-        mock_jwt.decode.return_value = {
-            "id": "1",
-            "credentials": {
-                "test-aws": {"encrypted_creds"},
-                "test-aws-cn": {"encrypted_creds"}
-            }
-        }
-        mock_decrypt.return_value = {
-            "access_key_id": "123456",
-            "secret_access_key": "654321"
-        }
-
-        job_id, credentials = self.service.decode_credentials(message)
-
-        mock_jwt.decode.assert_called_once_with(
-            'secret_credentials', 'super.secret', algorithm='HS256',
-            issuer='credentials', audience='obs'
-        )
-
-        assert len(credentials.keys()) == 2
-        assert job_id == '1'
-        assert credentials['test-aws']['access_key_id'] == '123456'
-        assert credentials['test-aws']['secret_access_key'] == '654321'
-
-        # Missing credentials key
-        mock_jwt.decode.return_value = {"id": "1"}
-
-        job_id, credentials = self.service.decode_credentials(message)
-        self.service.log.error.assert_called_once_with(
-            "Invalid credentials response recieved: 'credentials'"
-            " key must be in credentials message."
-        )
-
-        # Credential exception
-        self.service.log.error.reset_mock()
-        mock_jwt.decode.side_effect = Exception('Token is broken!')
-
-        job_id, credentials = self.service.decode_credentials(message)
-        self.service.log.error.assert_called_once_with(
-            'Invalid credentials response token: Token is broken!'
-        )
-
-    def test_decrypt_credentials(self):
-        self.service.encryption_keys_file = '../data/encryption_keys'
-        msg = b'gAAAAABaxoqn6i-IJAUaVXd6NkVqdJ8GKRiEDT9TgFkdS9r2U8NHyBoG' \
-            b'M2Bc4sUsTVBd1a3S7XCESxXgOdrTH5vUvj26TqkIuDTxg4lw-IIT3D84pT' \
-            b'6wX2cSEifMYIcjUzQGPXWhU4oQgrwOYIdR9p9DxTw5GPMwTQ=='
-
-        creds = self.service.decrypt_credentials(msg)
-
-        assert creds['access_key_id'] == '123456'
-        assert creds['secret_access_key'] == '654321'
-
-    @patch.object(MashService, 'get_credential_request')
-    @patch.object(MashService, '_publish')
-    def test_publish_credentials_request(
-        self, mock_publish, mock_get_credential_request
-    ):
-        token = Mock()
-        mock_get_credential_request.return_value = token
-
-        self.service.service_exchange = 'obs'
-        self.service.publish_credentials_request('1')
-
-        mock_publish.assert_called_once_with(
-            'credentials', 'request.obs', token
-        )
-
-    @patch.object(MashService, '_publish')
-    def test_publish_credentials_delete(self, mock_publish):
-        self.service.publish_credentials_delete('1')
-        mock_publish.assert_called_once_with(
-            'credentials',
-            'job_document',
-            JsonFormat.json_message({"credentials_job_delete": "1"})
-        )
-
-    @patch.object(MashService, '_publish')
-    def test_publish_credentials_delete_exception(self, mock_publish):
-        mock_publish.side_effect = AMQPError('Unable to connect to RabbitMQ.')
-
-        self.service.publish_credentials_delete('1')
-        self.service.log.warning.assert_called_once_with(
-            'Message not received: {0}'.format(
-                JsonFormat.json_message({"credentials_job_delete": "1"})
-            )
         )
 
     def test_should_notify(self):
